First commit of the local iPhone optimized changes.
Summary of modifications:

The original distribution was modified so that its zlib compatible optimized compressor can emit the default zlib header/trailer/CRC (this was the behavior before modifications), or optionally not emit the zlib header/trailer/CRC (this is called a 'raw zlib stream').  The optimized iPhone .png format requires a 'raw zlib stream' for the compressed IDAT chunk, whereas the PNG standard mandatated behavior is to include the zlib header/trailer/CRC.

An additional tool was added: `adviphone`.  After some trial and error, I went with this approach.  The `advpng` and `adviphone` tools do two different things.

The `advpng` tool was already present, and it *does not* work on iPhone .png files.  It does additional processing and decoding of the .png file that caused some difficulties in adapting it to read / write iPhone .png files.

What was really needed was a tool that did nothing more than decompress the .png's IDAT chunk and recompress it using the compression optimizing 7z zlib/deflat/rfc 1950 compression engine.  This preserves the .png filter settings that have likely already been optimized by a `pngcrush` pass.  The `advpng` tool does not do .png filter optimization, which is often a considerable source of compression gains.

Therefore, the `adviphone` tool was created.  The one and only thing it does is decompress the .png IDAT chunk and recompress it using the 7z compression engine.  This is vastly simpler and it doesn't require any special knowledge or handling of whatever h0h0 magical hacks Apple put in their Xcode `pngcrush` tool.  The .png file that comes out of `adviphone` is exactly the same as the original .png file EXCEPT for the IDAT chunk, and the only modification to that was that it was decompressed and recompressed, and that's all.  The decompressed IDAT bytes aren't examined, decoded, or interperted in any way.  Assuming there are no bugs in the 7z compressor (reasonable assumption), the decompressed bytes are identical to the originals.